refactor(ndt_scan_matcher): isolate implementation type to ndt
