+++ localization/ndt/include/ndt/base.hpp
@@ -22,6 +22,8 @@
 
 #include <vector>
 
+enum class NDTImplementType { PCL_GENERIC = 0, PCL_MODIFIED = 1, OMP = 2 };
+
 template <class PointSource, class PointTarget>
 class NormalDistributionsTransformBase
 {
@@ -32,6 +34,7 @@ public:
   virtual void align(pcl::PointCloud<PointSource> & output, const Eigen::Matrix4f & guess) = 0;
   virtual void setInputTarget(const pcl::shared_ptr<pcl::PointCloud<PointTarget>> & map_ptr) = 0;
   virtual void setInputSource(const pcl::shared_ptr<pcl::PointCloud<PointSource>> & scan_ptr) = 0;
+  virtual NDTImplementType getImplementationType() = 0;
 
   virtual void setMaximumIterations(int max_iter) = 0;
   virtual void setResolution(float res) = 0;
