+++ localization/ndt/include/ndt/pcl_modified.hpp
@@ -36,6 +36,7 @@ public:
   void align(pcl::PointCloud<PointSource> & output, const Eigen::Matrix4f & guess) override;
   void setInputTarget(const pcl::shared_ptr<pcl::PointCloud<PointTarget>> & map_ptr) override;
   void setInputSource(const pcl::shared_ptr<pcl::PointCloud<PointSource>> & scan_ptr) override;
+  NDTImplementType getImplementationType() override;
 
   void setMaximumIterations(int max_iter) override;
   void setResolution(float res) override;
