+++ localization/ndt/package.xml
@@ -15,6 +15,7 @@
   <depend>libpcl-all-dev</depend>
   <depend>ndt_omp</depend>
   <depend>ndt_pcl_modified</depend>
+  <depend>fmt</depend>
 
   <test_depend>ament_cmake_cppcheck</test_depend>
   <test_depend>ament_lint_auto</test_depend>
